fix; grad (integer) to grad (decimal / float)

+++ database/migrations/2024_05_24_174206_create_sheetmusics_table.php
@@ -21,8 +21,8 @@ return new class extends Migration
             $table->string('arranger');
             $table->string('publisher');
             $table->string('style');
-            $table->integer('grad')->default(0);
-            $table->integer('flex')->default(0);
+            $table->decimal('grad')->default(0);
+            $table->decimal('flex')->default(0);
             $table->time('duration')->default('00:00:00');
             $table->boolean('is_digital')->default(false);
             $table->string('demo_url')->nullable();
